Renamed files, Deleted the barrel file, Updated imports, added jsdocs
