+++ apps/backend/src/repositories/aggregation.repository.ts
@@ -23,6 +23,12 @@ interface UserSchoolRecord {
   schoolName: string;
 }
 
+/**
+ * Aggregation Repository
+ *
+ * Provides data access methods for data aggregation services upstream.
+ * Extends BaseRepository for standard CRUD operations.
+ */
 export class AggregationRepository {
   constructor(private readonly coreDb: NodePgDatabase<typeof CoreDbSchema> = CoreDbClient) {}
